+++ bin/src/api/auth/index.ts
