+++ bin/src/api/filesystem/index.ts
@@ -1,0 +1,3 @@
+export * from './models'
+export * from './file/file.controller';
+export * from './folder/folder.controller';
