+++ bin/src/api/filesystem/models/file.model.ts
