+++ bin/src/api/filesystem/models/folder.model.ts
