--- bin/src/filesystem/index.ts
@@ -1,4 +1,0 @@
-export * from './models/folder.model';
-export * from './models/file.model';
-export * from './folder/folder.controller';
-export * from './file/file.controller';
